BUGFIX: remove use of bfq queues after free
bfq queues occasionally happened to be used after being freed, because
they were accessed after some invocations of bfq_put_queue that could
cause them to be freed. This commit refactors code, when needed, to
avoid any occurrence of such a use-after-free of a bfq queue. This
commit also adds comments to make references to bfq queues easier to
follow.

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: blinoff82 <blinov.in@gmail.com>